backend/requirements.txt
asn1crypto==0.24.0
astroid==1.6.5
attrs==18.1.0
Automat==0.7.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.3
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
configparser==3.5.0
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
django-cors-headers==2.4.0
enum34==1.1.6
functools32==3.2.3.post2
futures==3.2.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipaddress==1.0.22
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.4
mccabe==0.6.1
parsel==1.5.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==1.9.3
pyOpenSSL==18.0.0
queuelib==1.5.0
requests==2.19.1
Scrapy==1.5.1
scrapy-splash==0.7.2
service-identity==17.0.0
singledispatch==3.4.0.3
six==1.11.0
Twisted==18.7.0
urllib3==1.23
virtualenv==16.0.0
w3lib==1.19.0
webencodings==0.5.1
wrapt==1.10.11
zope.interface==4.5.0
